Fixing scripts

+++ src/aws-deploy.sh
@@ -15,11 +15,6 @@ if [ -z "$SERVICE_NAME" ]; then
     exit -1
 fi
 
-if [ -z "$SRC_JAR" ]; then
-    echo "ERROR: Please, set up SRC_JAR (path to source app jar) variable before running this script"
-    exit -1
-fi
-
 if [ -z "$SRC_PROPS" ]; then
     echo "ERROR: Please, set up SRC_PROPS (path to source app properties) variable before running this script"
     exit -1
@@ -40,13 +35,19 @@ if [ -z "$AWS_LOGIN" ]; then
     exit -1
 fi
 
+if [ -z "$SRC_JAR" ]; then
+    echo "INFO: Source jar is missing, skipping jar checks and copy"
+fi
+
 #
 # Check for existence of source files
 #
 
-if [ ! -f $SRC_JAR ]; then
-    echo "ERROR: No jar file at $SRC_JAR"
-    exit -1
+if [ ! -z "$SRC_JAR" ]; then
+    if [ ! -f $SRC_JAR ]; then
+        echo "ERROR: No jar file at $SRC_JAR"
+        exit -1
+    fi
 fi
 
 if [ ! -f $SRC_PROPS ]; then
@@ -67,7 +68,12 @@ echo "[OK] Initial checks"
 
 RAND_PREFIX=$(xxd -l 10 -p /dev/random)
 
-TARGET_JAR=/tmp/$RAND_PREFIX-app.jar
+if [ ! -z "$SRC_JAR" ]; then
+    TARGET_JAR=/tmp/$RAND_PREFIX-app.jar
+else
+    TARGET_JAR=
+fi
+
 TARGET_PROPS=/tmp/$RAND_PREFIX-app.properties
 
 TEMP_LAYOUT_SCRIPT=/tmp/$RAND_PREFIX-setup-layout.sh
@@ -77,8 +83,10 @@ TEMP_SERVER_SCRIPT=/tmp/$RAND_PREFIX-server.sh
 # Deploy and restart
 #
 
-scp -i $AWS_PEM $SRC_JAR $AWS_LOGIN@$AWS_HOSTNAME:$TARGET_JAR
-echo "[OK] Copied $SRC_JAR to $TARGET_JAR"
+if [ ! -z "$SRC_JAR" ]; then
+    scp -i $AWS_PEM $SRC_JAR $AWS_LOGIN@$AWS_HOSTNAME:$TARGET_JAR
+    echo "[OK] Copied $SRC_JAR to $TARGET_JAR"
+fi
 
 scp -i $AWS_PEM $SRC_PROPS $AWS_LOGIN@$AWS_HOSTNAME:$TARGET_PROPS
 echo "[OK] Copied $SRC_PROPS to $TARGET_PROPS"
@@ -95,6 +103,7 @@ echo "[OK] Temporary directory structure script created at $TEMP_LAYOUT_SCRIPT"
 # Create contents of TEMP_SERVER_SCRIPT
 echo "
 export SERVICE_NAME=$SERVICE_NAME
+export SERVER_START_ACTION=restart
 " > $TEMP_SERVER_SCRIPT
 cat server.sh >> $TEMP_SERVER_SCRIPT
 echo "[OK] Temporary server script created at $TEMP_SERVER_SCRIPT"
@@ -108,6 +117,8 @@ echo "[OK] Temporary server script executed, server should be restarted"
 
 # Cleanup
 
-rm -rf /tmp/${RAND_PREFIX}*
-ssh -i $AWS_PEM $AWS_LOGIN@$AWS_HOSTNAME "rm -rf /tmp/${RAND_PREFIX}*"
-echo "[OK] Cleanup done"
+if [ -z "$KEEP_TEMP_FILES" ]; then
+    rm -rf /tmp/${RAND_PREFIX}*
+    ssh -i $AWS_PEM $AWS_LOGIN@$AWS_HOSTNAME "rm -rf /tmp/${RAND_PREFIX}*"
+    echo "[OK] Cleanup done"
+fi
